feat: global search api

+++ BuildingBlocks/Core/BuildingBlocks.Application/Dtos/PagedResultDto.cs
@@ -2,11 +2,11 @@
 
 public class PagedResultDto<T>
 {
-    public int Total { get; set; }
+    public long Total { get; set; }
     public int TotalPages { get; set; }
     public IEnumerable<T> Data { get; set; }
     
-    public PagedResultDto(int total, int size, IEnumerable<T> data)
+    public PagedResultDto(long total, int size, IEnumerable<T> data)
     {
         Total = total;
         TotalPages = (int)Math.Ceiling(total * 1.0 / size);
